requirements.txt
pandas>=2.0.0
selenium>=4.15.0
undetected-chromedriver>=3.5.4
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
tqdm>=4.66.0